Refactor CsvReader namespace
